refactor: rename InputNumber->AppNumber
